requirements.txt
numpy>=1.26.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
nltk>=3.8.0
tqdm>=4.65.0
openai>=1.0.0
anthropic>=0.3.0
git+https://github.com/csebuetnlp/normalizer
bnlp_toolkit>=4.0.0
sentencepiece>=0.2.0

